from double precision to single precision

+++ more_work.cl
@@ -1,10 +1,10 @@
 
 __kernel //__attribute__ ((reqd_work_group_size(2048, 2048, 1)))
-void more_work(__global double* a, __global double* b, __global double* output)
+void more_work(__global float* a, __global float* b, __global float* output)
 {
     int r = get_global_id(0);
     int c, index;
-    double running;
+    float running;
     int rank = get_global_size(0);
 
     for (c=0; c < rank; c++) {
